# perf: fs optimizations - part 1