--- a/runtime/js/30_fs.js
+++ b/runtime/js/30_fs.js
@@ -205,9 +205,7 @@
           str += `${name}: view[${offset}] + view[${offset + 1}] * 2**32 ,`;
         } else {
           str +=
-            `${name}: view[${offset}] === 0 && view[${offset} + 1] === 0 && view[${
-              offset + 1
-            }] === 0 ? null : new Date(view[${offset}] + view[${
+            `${name}: view[${offset}] === 0 ? null : new Date(view[${offset}] + view[${
               offset + 1
             }] * 2**32),`;
         }