--- a/runtime/js/30_fs.js
+++ b/runtime/js/30_fs.js
@@ -198,9 +198,9 @@
         if (!optional) {
           str += `${name}: view[${offset}] + view[${offset + 1}] * 2**32,`;
         } else {
-          str += `${name}: (view[${offset}] + view[${
-            offset + 1
-          }] * 2**32) || null ,`;
+          str += `${name}: (view[${offset}] + view[${offset + 1}] * 2**32) ${
+            Deno.build.os === "windows" ? "|| null" : ""
+          },`;
         }
       } else if (type == "date") {
         str +=