--- a/runtime/js/30_fs.js
+++ b/runtime/js/30_fs.js
@@ -265,9 +265,9 @@
       isDirectory: response.isDirectory,
       isSymlink: response.isSymlink,
       size: response.size,
-      mtime: response.mtimeSet != null ? new Date(response.mtime) : null,
-      atime: response.atimeSet != null ? new Date(response.atime) : null,
-      birthtime: response.birthtimeSet != null
+      mtime: response.mtimeSet !== null ? new Date(response.mtime) : null,
+      atime: response.atimeSet !== null ? new Date(response.atime) : null,
+      birthtime: response.birthtimeSet !== null
         ? new Date(response.birthtime)
         : null,
       // Only non-null if on Unix