- [x] Flatten op arguments.
  
  As a result, `truncateSync` becomes 2.3x faster.
  ```
  # main
  () => truncateSync("test", 0)
  time 1029 ms rate 97181
  time 1013 ms rate 98716
  time 745 ms rate 134228
  
  # This patch
  () => truncateSync("test", 0)
  time 469 ms rate 213219
  time 447 ms rate 223713
  ```
- [x] ~8x faster way to serialize `op_stat_sync` result to JS.

  ```
  # main
  () => lstatSync("test")
  time 767 ms rate 130378
  time 767 ms rate 130378
  time 745 ms rate 134228
  
  # This patch
  () => lstatSync("test")
  time 99 ms rate 1010101
  time 99 ms rate 1010101
  time 98 ms rate 1020408
  ```
- [x] ~14x faster file copy on MacOS using APFS' `clonefile`

  ```
  # main
  () => copyFileSync("test", "test2")
  time 8646 ms rate 1156
  time 7578 ms rate 1319

  # This patch
  () => copyFileSync("test", "test2")
  time 738 ms rate 13550
  time 647 ms rate 15455
  ```
- [x] Fast path for `perms.read.check` / `perms.write.check` to avoid `resolve_from_cwd` when possible. It's quite literally a bool check when run with `-A`, making it infintely faster than before.